fix(release): ensure debuginfo file contains debugging info
Seems recently Cargo defaulted to use `strip=debuginfo`. This
inadvertently made our debuginfo files much smaller

Fix the issue by using `strip=none` and add a test so that it breaks if
this somehow changes again.
